// IMFS-Web-FrontEnd/src/app/models/application/application.model.ts
export class RecentApplications {
    status: string;
    message: string;
    recentAppDetails: RecentApplicationsModel[];
}

export class RecentApplicationsModel {
    id?: number;
    applicationNumber?: number;
    endUser: string;
    status: string;
    finanaceAmount?: number;
    financeType?: string;
    createdDate?: Date;
}

export class AwaitingInvoices {
    status: string;
    message: string;
    recentInvoiceDetails: AwaitingInvoiceModel[];
}

export class AwaitingInvoiceModel {
    id?: number;
    applicationNumber?: number;
    status: string;
    endCustomerName: string;
    finanaceAmount?: number;
    financeType?: string;
    createdDate?: Date;
    approvedDate?: Date;
}